+++ cmd/aqt/main.go
@@ -61,8 +61,9 @@ var (
 )
 
 func main() {
+	argv := multiCallArgs(os.Args)
 	root := rootCmd()
-	root.SetArgs(escapeLeadingDashIDs(root, os.Args[1:]))
+	root.SetArgs(escapeLeadingDashIDs(root, argv[1:]))
 	cmd, err := root.ExecuteC()
 	if err != nil {
 		fmt.Fprintln(os.Stderr, "error:", explainError(err))
@@ -294,7 +295,7 @@ func rootCmd() *cobra.Command {
 	root.AddCommand(initCmd(), statusCmd(), diffCmd(), syncCmd(), cloneCmd(), watchCmd(), agentCmd())
 	root.AddCommand(snapshotCmd(), checkpointCmd(), restoreCmd())
 	root.AddCommand(sharesCmd(), contactsCmd())
-	root.AddCommand(repoCmd())
+	root.AddCommand(repoCmd(), gitCmd())
 	root.AddCommand(gitRemoteHelperCmd())
 	root.AddCommand(tuiCmd(), updateCmd())